Fix app indicator when using custom install prefix
When Remmina is installed to a custom location (e.g. when using
'cmake -DCMAKE_INSTALL_PREFIX=/opt/remmina') the app indicator applet in
Unity doesn't display because it tries to load the icon from one of the
standard icon theme locations (e.g. /usr/share/icons, etc). When it
doesn't find the Remmina icon, the app indicator is not displayed in the
notification panel. With this commit, the app indicator is configured to
look in the correct location for the icon. As a result, the app
indicator properly loads into the notification panel.