rdp-plugin: Advertise gfx:rfx correctly
With the introduction of the graphics pipeline in Windows 8, RDP
supports two types of the RemoteFX codec.
First is the RemoteFX Progressive codec, advertised by just
setting the SupportGraphicsPipeline to TRUE, where the data is
transfered via the RDPGFX_WIRE_TO_SURFACE_PDU_2 pdu and the second one
is the "normal" RemoteFX codec, which also existed for Windows 7 with
the difference that it is now transfered via the graphics pipeline
via the RDPGFX_WIRE_TO_SURFACE_PDU_1 pdu.

Remmina, however, currently doesn't activate the RemoteFXCodec when
using gfx:rfx.
Fix this error by advertising the RemoteFX codec correctly.