rdp-plugin: Add new color depth setting "Automatic"
Currently, Remmina assumes for each connection that the server always
supports H264 and attempts to use it, if FreeRDP is built with H264
support.
However, in reality not every server supports H264 which will have the
effect, since Remmina won't advertise any other codec simultaneously to
the H264 codec, that the server will fall back to sending raw bitmaps,
since it does not support H264, which results in a massive performance
regression (including massive bandwidth usage), although the server
can still support other codecs.
This raises the users thought that this is actually the server's fault,
although the server can choose a better codec, if Remmina would
advertise it.
mstsc (the RDP client in MS Windows) solves this situation by
advertising all codecs it supports by default.

Do the same in Remmina to make sure that the server doesn't fall back
to the worst transfer method unless the user forces a different codec
in the RDP colour setting.
This fixes an unwanted performance regression in both Weston's RDP
backend and gnome-remote-desktop's RDP backend which does not happen in
mstsc.