rdp-graphics: Remove wrong NULL check on andMaskData
Remmina currently assumes that both the xorMaskData and the andMaskData
have to be non-NULL values in order to be able to process the new
mouse pointer bitmap.
This, however, is not entirely true, as the andMaskData is allowed to be
NULL, especially when there is no need for the andMask, e.g. for 32-bit
mouse pointer bitmaps.
freerdp_image_copy_from_pointer_data(), the function that both xfreerdp
and Remmina use, also processes the andMask, if it is non-NULL.
The same applies to mstsc (the RDP client MS Windows).
The current NULL check has the effect, that Remmina falls into a
connect-disconnect loop when the server sends a new pointer bitmap
with a NULL andMask (because it is not necessary).
This, however, works completely fine in both mstsc and xfreerdp.
Only Remmina falls into this trap, making the RDP connection unusable.

Therefore, fix this behaviour by removing the NULL check for the
andMaskData, as the current check is wrong and
freerdp_image_copy_from_pointer_data() will already return FALSE, when
the pointer data is not valid.
This fixes Remmina not being usable with the RDP backend in
gnome-remote-desktop.