[RFC][v2] rdp: add Use base credential for RD gateway autentication
Even if rfi->settings->GatewayUseSameCredentials is TRUE (setted when gw
user/password/domain are empty) remmina still request gw credentials in some
cases.
This because GatewayUseSameCredentials in freerdp use gateway credentials also
for base, as opposite of remmina.

This patch add a flag for choose is use base credentials also for gateway
credentials.
In remmina_rdp_gw_authenticate use and save base cred if the new option is true.
In remmina_rdp_main reset gateway credentials if the new option is true.
Keeped for retrocompability the code for "auto" use base settings for gateway
without this flag (that was working for the case where both base username and
password are stored)

I want ask a question:
I keeped for retrocompability the code about "auto" use base settings for
gateway (that was working for the case where both base username and password are
stored), or with new flag is not good?

Any other advice for improve the patch is appreciated.