rdp/cliprdr: Fix header of FormatList message
Currently, Remmina does not set the msgType and sets the wrong msgFlag
for the FormatList message.

The clipboard documentation of RDP states:

The msgType field of the Clipboard PDU Header MUST be set to
CB_FORMAT_LIST (0x0002), while the msgFlags field MUST be set to
0x0000 or CB_ASCII_NAMES (0x0004) depending on the type of data present
in the formatListData field.

Fix this issue by setting the correct msgType and msgFlag.