rdp/event: Fix wheel value for GDK_SCROLL_DOWN events
For GDK_SCROLL_DOWN events, the negative wheel value needs to be
processed by the server.
However, instead of submitting the value in the two's complement,
Remmina needs to submit the absolute value in addition to the
PTRFLAGS_WHEEL_NEGATIVE flag.

Quoting akallabeth from IRC (#freerdp) here:
"0 to 0xFF is the value range, 0x100 the sign bit"

Fix this issue by using the PTRFLAGS_WHEEL_NEGATIVE flag in addition
to the absolute value.
This has the effect that mstsc, xfreerdp and Remmina all produce the
same correct wheel value on the server side when scrolling down.