Implementing simple SSH multi factor authentication.
When an SSH server AuthenticationMethods with multiple methods, like
public key authentions and password, we receive an SSH_PARTIAL_AUTHENTICATION
return code, that was ignored in Remmina.

Now Remmina correctly handle this state and pass to the second mechanism.

More than 2 mechanism are not supported yet, because Remmina doesn't go
behind 3 authentions attempts at the moment (to be fixed).

The next step will be to implement keyboard interactive authentication.

Signed-off-by: Antenore Gatta <antenore@simbiosi.org>