Fixing RemminaConnectionWindow map/unmap events
The remmina connection windows in fullscreen is a different object
than when is in scrolled mode, therefore the old logic to map and unmap
the windows from the plugin it was notr working.

Now the events are managed directly from the RCW object and the
RemminaProtocolWidget, with an API that can be used by any plugins.

When in multi monitor and fullscreen, this mechanism doesn't work,
Therefore, in the plugin itself I catch when we are in multi monitor fullscreen mode
And I ignore the event.

This last behaviour should be managed directly in the Remmina Protocol Widget (TODO).

Fixes #2475

Signed-off-by: Antenore Gatta <antenore@simbiosi.org>