Adding option to disable smooth scrolling
Smooth scrolling is used adding the GDK_SMOOTH_SCROLL_MASK bit-flag to the remmina drawing area.
The event is managed at a plugin level, at the moment only for RDP and VNC.

In RDP I have added a genral RDP settings that apply too all profiles, and a setting to disable
smooth scrolling on a connection profile level.

In this way is possible to disable globally the smooth scrolling, or on a per connection profile.

In VNC, as it doesn't have global settings, I've added only on a per connection level.

Closes #2531

Signed-off-by: Antenore Gatta <antenore@simbiosi.org>