Patch for a Remmina segfault
This is a "-p1" patch for Remmina 1.4.22 and  "git" snapshots from the
same period.  It's  a fix for a segfault that shouldn't  introduce new
problems.

One issue with the patch is that it doesn't re-indent the code between
the "{" and "}" added here.  It was done this way to keep the patch as
simple as possible.

Signed-off-by: Robert Kiraly <me@boldcoder.com>